Patched code to build with FLAC on Ubuntu 12.04
This is mostly based on this patch, which also enabled Windows x86-64 support:
https://bugs.launchpad.net/ubuntu/+source/audiere/+bug/234900

However I have also made some extra tweaks to fix a remaining bug.

+++ src/input_flac.h
@@ -37,7 +37,7 @@ namespace audiere {
     static FLAC__StreamDecoderReadStatus read_callback(
       const FLAC__StreamDecoder* decoder,
       FLAC__byte buffer[],
-      size_t* bytes,
+      long unsigned int* bytes,
       void* client_data);
     static FLAC__StreamDecoderSeekStatus seek_callback(
       const FLAC__StreamDecoder* decoder,
